Do not leave an .svn file behind when running the tests.
Do create a file, but call it `.dummy` for good measure, and remove it
afterwards.

The `.svn` was presumably from the days when plone.resource was still
in subversion, which for older versions of subversion meant that there
was a `.svn` directory already there, and should definitely not be
removed.