Fixed 'ValueError: substring not found' in ``FilesystemResourceDirectory`` representation.
This happens when you register a directory with a name that differs from the directory name.
Visiting the ``/++theme++myname`` url would then give this error.

We also avoid listing a longer part of the path in case the directory name happens to be in the path multiple times.
Case in point: if you would register the `template` dir of `plone.app.theming` as a plone static dir with name `theming`, and visit the `++theme++theming` url,
you would see something like this:

  <FilesystemResourceDirectory object at theming-1.1.8-py2.7.egg/plone/app/theming/themes/template>

That is just silly.
We simply show the name now:

  <FilesystemResourceDirectory object at theming>

If the name and the directory would have matched, which is the default when you do not give a name, then that would be what is shown anyway.